requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.10.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=0.27.0
boto3>=1.26.0
pytest>=7.3.1
httpx>=0.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
python-decouple>=3.8
jinja2>=3.0.0
